refactor(appframe): drop cjs; convert to esm only
BREAKING CHANGE: remove cjs build
BREAKING CHANGE: remove file imports (eg, 'packagename/react')
