+++ packages/appframe/src/index.ts
@@ -1,3 +1,3 @@
-export { default } from './react'
-export * from './vars'
-export { default as css } from './css'
+export { default } from './react/index.js'
+export * from './vars/index.js'
+export { default as css } from './css/index.js'
