+++ packages/appframe/src/react/__specs__/index.spec.tsx
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react'
 import React from 'react'
 
-import AppFrame from '..'
+import AppFrame from '../index.js'
 
 describe('AppFrame', () => {
   it('renders', () => {
